--- components/layout/Footer.js
@@ -1,12 +1,0 @@
-
-function Footer() {
-  return (
-    <div className={"flex flex-col items-center justify-center bg-black p-10"}>
-      <p className={"text-white"}>Developed by: Alejandro Lopez</p>
-      <p className={"text-white"}>Text here</p>
-      <p className={"text-white"}>Text here</p>
-    </div>
-  );
-}
-
-export default Footer;
