+++ components/layout/Footer.tsx
@@ -1,0 +1,10 @@
+
+function Footer() {
+  return (
+    <div className={"flex flex-col items-center justify-center bg-black py-10"}>
+      <p className={"text-white"}>Developed by: Alejandro Lopez</p>
+    </div>
+  );
+}
+
+export default Footer;
